requirements.txt
altair==5.1.2
annotated-types==0.6.0
anyio @ file:///work/ci_py311/anyio_1676823771847/work/dist
argon2-cffi @ file:///opt/conda/conda-bld/argon2-cffi_1645000214183/work
argon2-cffi-bindings @ file:///work/ci_py311/argon2-cffi-bindings_1676823553406/work
asttokens @ file:///opt/conda/conda-bld/asttokens_1646925590279/work
attrs @ file:///work/ci_py311/attrs_1676822758931/work
backcall @ file:///home/ktietz/src/ci/backcall_1611930011877/work
beautifulsoup4 @ file:///croot/beautifulsoup4-split_1681493039619/work
bleach @ file:///opt/conda/conda-bld/bleach_1641577558959/work
catalogue==2.0.10
certifi==2023.7.22
cffi @ file:///work/ci_py311/cffi_1676822533496/work
cfgv==3.4.0
charset-normalizer==3.2.0
click==8.1.7
cmake==3.27.4.1
comm @ file:///work/ci_py311/comm_1677709131612/work
confection==0.1.3
contourpy==1.1.1
cycler==0.11.0
debugpy @ file:///croot/debugpy_1690905042057/work
decorator @ file:///opt/conda/conda-bld/decorator_1643638310831/work
defusedxml @ file:///tmp/build/80754af9/defusedxml_1615228127516/work
dill==0.3.7
distlib==0.3.7
entrypoints @ file:///work/ci_py311/entrypoints_1676823319002/work
executing @ file:///opt/conda/conda-bld/executing_1646925071911/work
fastcore==1.5.29
fastjsonschema @ file:///work/ci_py311_2/python-fastjsonschema_1679340124475/work
filelock==3.12.3
fonttools==4.42.1
fuzzywuzzy==0.18.0
identify==2.5.31
idna @ file:///work/ci_py311/idna_1676822698822/work
imageio==2.31.6
imgaug==0.4.0
ipykernel @ file:///croot/ipykernel_1691121631942/work
ipython @ file:///croot/ipython_1691532092695/work
ipython-genutils @ file:///tmp/build/80754af9/ipython_genutils_1606773439826/work
jedi @ file:///work/ci_py311_2/jedi_1679336495545/work
Jinja2 @ file:///work/ci_py311/jinja2_1676823587943/work
joblib==1.3.2
jsonlines==4.0.0
jsonschema @ file:///croot/jsonschema_1678983421637/work
jupyter-server @ file:///work/ci_py311/jupyter_server_1676824947163/work
jupyter_client @ file:///work/ci_py311/jupyter_client_1676823491351/work
jupyter_core @ file:///croot/jupyter_core_1679906564508/work
jupyterlab-pygments @ file:///tmp/build/80754af9/jupyterlab_pygments_1601490720602/work
kiwisolver==1.4.5
lazy_loader==0.3
Levenshtein==0.23.0
lit==16.0.6
loguru==0.7.2
lxml @ file:///croot/lxml_1679646450407/work
markdown-it-py==3.0.0
MarkupSafe @ file:///work/ci_py311/markupsafe_1676823507015/work
matplotlib==3.8.0
matplotlib-inline @ file:///work/ci_py311/matplotlib-inline_1676823841154/work
mdurl==0.1.2
mistune @ file:///work/ci_py311/mistune_1676823698136/work
mpmath==1.3.0
nb-conda-kernels @ file:///work/ci_py311/nb_conda_kernels_1676845026630/work
nbclassic @ file:///croot/nbclassic_1681756170010/work
nbclient @ file:///work/ci_py311/nbclient_1676823713954/work
nbconvert @ file:///work/ci_py311/nbconvert_1676823854564/work
nbformat @ file:///work/ci_py311/nbformat_1676823523130/work
nest-asyncio @ file:///work/ci_py311/nest-asyncio_1676823382924/work
networkx==3.1
nltk==3.8.1
nodeenv==1.8.0
notebook @ file:///croot/notebook_1690984815942/work
notebook_shim @ file:///work/ci_py311/notebook-shim_1676825053954/work
numpy==1.26.0
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
opencv-python==4.8.0.76
packaging @ file:///croot/packaging_1693575174725/work
pandas==2.1.1
pandocfilters @ file:///opt/conda/conda-bld/pandocfilters_1643405455980/work
parso @ file:///opt/conda/conda-bld/parso_1641458642106/work
pexpect @ file:///tmp/build/80754af9/pexpect_1605563209008/work
pickleshare @ file:///tmp/build/80754af9/pickleshare_1606932040724/work
Pillow==10.0.1
platformdirs @ file:///croot/platformdirs_1692205439124/work
pre-commit==3.5.0
prometheus-client @ file:///work/ci_py311_2/prometheus_client_1679340520168/work
prompt-toolkit @ file:///work/ci_py311/prompt-toolkit_1676824857154/work
psutil @ file:///work/ci_py311_2/psutil_1679337388738/work
ptyprocess @ file:///tmp/build/80754af9/ptyprocess_1609355006118/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl
pure-eval @ file:///opt/conda/conda-bld/pure_eval_1646925070566/work
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pydantic==2.4.2
pydantic_core==2.10.1
Pygments @ file:///croot/pygments_1684279966437/work
pyhere==1.0.0
PyMuPDF==1.23.5
PyMuPDFb==1.23.5
pyparsing==3.1.1
pyrsistent @ file:///work/ci_py311/pyrsistent_1676823088752/work
python-dateutil @ file:///tmp/build/80754af9/python-dateutil_1626374649649/work
python-Levenshtein==0.23.0
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq @ file:///work/ci_py311/pyzmq_1676918885689/work
rapidfuzz==3.4.0
regex==2023.10.3
requests==2.31.0
rich==13.6.0
scikit-image==0.22.0
scikit-learn==1.3.1
scipy==1.11.2
Send2Trash @ file:///tmp/build/80754af9/send2trash_1632406701022/work
shapely==2.0.2
six @ file:///tmp/build/80754af9/six_1644875935023/work
sniffio @ file:///work/ci_py311/sniffio_1676823742530/work
soupsieve @ file:///croot/soupsieve_1680518478486/work
srsly==2.4.8
stack-data @ file:///opt/conda/conda-bld/stack_data_1646927590127/work
sympy==1.12
terminado @ file:///work/ci_py311/terminado_1677696151350/work
threadpoolctl==3.2.0
tifffile==2023.9.26
tinycss2 @ file:///work/ci_py311/tinycss2_1676823757641/work
toolz==0.12.0
torch==2.0.1
torch-snippets==0.507
torchaudio==2.0.2
torchsummary==1.5.1
torchvision==0.15.2
tornado @ file:///croot/tornado_1690848263220/work
tqdm==4.66.1
traitlets @ file:///work/ci_py311/traitlets_1676823305040/work
triton==2.0.0
typing==3.7.4.3
typing_extensions @ file:///croot/typing_extensions_1690297465030/work
tzdata==2023.3
urllib3==2.0.4
virtualenv==20.24.6
wasabi==1.1.2
wcwidth @ file:///Users/ktietz/demo/mc3/conda-bld/wcwidth_1629357192024/work
webencodings==0.5.1
websocket-client @ file:///work/ci_py311/websocket-client_1676824890004/work
xmltodict==0.13.0
